requirements.txt
# Audio
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4

# STT
openai-whisper==20231117

# Base de datos
psycopg2-binary==2.9.10

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
slowapi==0.1.9
bcrypt==4.2.1

# Comunicación serie
pyserial==3.5

# Configuración
pyyaml==6.0.1
python-dotenv==1.0.1

# Telegram
requests==2.31.0
